Error with Iris and Glass
find_attribute_probability uses the value of the record (ex 1-10 for BCD) as a way to index an array, so for the data with a high numbers in the records, this is a problem. I put a TODO comment where I need some help. It works in its current state because I made sure it wouldn't error out, but it needs to be fixed.